move history related functions to history.py
